# torch.compile() support