This PR adds support for compiling models in vLLM with torch.compile().

Consider this as a first draft/MVP. I've only tested it with Llama 7B-F in `examples/offline_inference.py`, and the results match when running with `enforce_eager=True/False`.

For this particular case, model compilation/warmup takes 151s + 12s from cudagraphs.

Next step is running some benchmarks to see if there's any improvements in timings.