@ani300 2024-02-28
These are benchmark results for latency running on the AWS p4de instances I have access to:

`python benchmarks/benchmark_latency.py --model=_llama_7b_ --input-len 1000 --output-len 50 --num-iters 100 --batch-size {1.2^n.64}`

main:
BS 1: 0.6218338459543884 s
BS 2: 0.6896187016926706 s
BS 4: 0.8732671845890582 s
BS 8: 1.2548897478543222 s
BS 16: 1.987721082046628 s
BS 32: 3.4572774835303424 s
BS 64: 6.404235514998436 s

this branch:
BS 1: 0.6227266596630215 s
BS 2: 0.6994614476896822 s
BS 4: 0.8837944919243454 s
BS 8: 1.2816928111203014 s
BS 16: 2.0030816675908865 s
BS 32: 3.5334619791805744 s
BS 64: 6.515135741680861 s

As I expected, the timings are pretty similar, with compile being on average ~1.37% slower. The explanation for this is two-fold:
1. torch.compile() adds a very small CPU overhead at the beginning of execution due to a cache lookup for the right compiled model out of all the ones getting compiled.
2. torch.compile() cannot further optimize all the custom CUDA kernels in the vLLM codebase. The power of compile lies in creating equivalent triton kernels to some of these custom CUDA kernels without the need to write C++ code or depend on a certain ABI. Given that most of the forward pass code of a model like Llama in vLLM is custom CUDA kernels at this point, we don't get the advantages of compile.