@sh1ng 2024-03-01
I also don't see any performance improvements https://github.com/vllm-project/vllm/pull/2131.
Probably vllm is already very optimized compared to vanilla torch.